Move AssignTesterDropdown
